+++ client/src/views/Home/HomePage.vue
@@ -1,0 +1,6 @@
+<template>
+  <h2>Home Page</h2>
+  <button @click="logout">Logout</button>
+</template>
+
+<script src="./Home.js" />
