+++ client/src/views/Landing/Landing.js
@@ -1,0 +1,8 @@
+export default {
+  name: "LandingPage",
+  methods: {
+    addExpense() {
+      this.$router.push("/login");
+    },
+  },
+};
