+++ client/src/views/Landing/LandingPage.vue
@@ -1,0 +1,6 @@
+<template>
+  <h2>Landing Page</h2>
+  <button @click="addExpense">Add expenses</button>
+</template>
+
+<script src="./Landing.js" />
